tidb/tidb-python-mysqlconnector-quickstart/requirements.txt
python-dotenv==1.0.0
mysql-connector-python>=8.0,<8.1